Update README.md
Intro to if statements

Examine the given program in the main.cpp code editor panel.

Run the program and see what happens when entering 10 or a different number.

Modify the program so that it will output a message when a user guesses your favorite number. For example, if your favorite number is 17 your program could have a sample run like this:

Enter a number:
17
Wow 17 is my favorite number!